+++ internal/router/contact_routes.go
@@ -11,4 +11,5 @@ func initContactRoutes(r fiber.Router, h *handler.Handler, authMw *middleware.Au
 	contactGroup.Use(authMw.JWTAuthentication())
 	contactGroup.Get("/check", h.WhatsappMessageHandler.CheckNumber)
 	contactGroup.Get("/", h.WhatsappMessageHandler.ListContacts)
+	contactGroup.Get("/info", h.WhatsappMessageHandler.GetContactInfo)
 }